Location is done

+++ INNOEcoSystem.Domain/Entities/Locations/Location.cs
@@ -5,8 +5,8 @@ namespace INNOEcoSystem.Domain.Entities.Locations;
 
 public class Location : Auditable
 {
-    public long LongiTude { get; set; }
-    public long Latitude { get; set; }
+    public decimal LongiTude { get; set; }
+    public decimal Latitude { get; set; }
     public string Addres { get; set; }
 
     public ICollection<LocationAsset> Assets { get; set; }
